requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
django==5.2.3
django-anymail==13.0
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-unfold==0.60.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
hiredis==3.2.1
idna==3.10
kombu==5.5.4
markdown==3.8.2
packaging==25.0
pillow==11.3.0
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pyjwt==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
ruff==0.12.7
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
